Factor out app.tsx, app-fader.tsx, and style.ts.
